Umstellung von Double auf Decimal sowie Complex auf DecComplex

+++ JScience/Physik/Simulationen/Wavefunctions/Hamiltonoperators/Potentials/Interfaces/IPotential.cs
@@ -6,6 +6,6 @@ namespace JScience.Physik.Simulationen.Wavefunctions.Hamiltonoperators.Potential
     public interface IPotential<T> : IHamilton<T> where T : IWavefunction
     {
         public string Name { get; }
-        public double Potential { get; }
+        public decimal Potential { get; }
     }
 }
