+++ JScience/Physik/Simulationen/Wavefunctions/TimeEvolution/Interfaces/IU_T.cs
@@ -6,7 +6,7 @@ namespace JScience.Physik.Simulationen.Wavefunctions.TimeEvolution.Interfaces
 {
     public interface IU_T<T> where T : IWavefunction
     {
-        double t_STEP { get; }
+        decimal t_STEP { get; }
 
         T Do(T WF, List<IHamilton<T>> Hamiltons);
     }
